Improvement flow batches doc
