+++ spec/tags/flows-batches.md
@@ -1,9 +1,6 @@
-The flows-batches api makes it possible to send a message with one or more contents to multiple contacts at once.
-<br/><br/>
-Use your flowId.
+The flows-batches api makes it possible to dispatch a active Flow with one or more contents to multiple contacts at once.
 
-<br/>
-
+To perform these operations, you will need your flowId and you can find it <a href="https://app.zenvia.com/flows" target="_blank">here</a>.
 ### What you need to know about the contacts file
 
 * The file must be in **.csv** format;
